adjusted data handling TODO add block based dma to solve producer (adc) and consumer (UI) mismatch

+++ IoTM_/STM32CubeIDE/App/adc/adc.cpp
@@ -50,7 +50,7 @@ uint16_t* AdcDma::getValues() {
 	return mDmaBuffer;
 }
 
-float AdcDma::getChannelValue(uint8_t ch) const{
+uint16_t AdcDma::getChannelValue(uint8_t ch) const{
 	return mDmaBuffer[ch];
 }
